[Build] Add transitive linking from Python bridge
Tangentially related to #654.

When a consuming project links to `openassetio-python-bridge` they also
need to link `openassetio-core` (at minimum for the `LoggerInterface`).
Previously this had to be done explicitly by the consuming library,
since the `openassetio-core` dependency was `PRIVATE`.

So make `openassetio-core` a `PUBLIC` dependency of
`openassetio-python-bridge` to allow for the transitive dependency to
work. This makes sense since the core library is not intentionally
"hidden" within `openassetio-python-bridge`.

Signed-off-by: David Feltell <david.feltell@foundry.com>