[Build] Link correct targets for Python bridge
Part of #654. For linking `openassetio-python-bridge` we were relying on
dodgy behaviour from the ConanCenter pybind11 package, which would add
Python linker flags even for the header-only `pybind11::pybind11`
target.

If/when this is fixed, we can find ourselves with "undefined reference"
errors (apparently only on macOS and Windows, Linux builds fine).

So update to use the correct targets. Note that even though
`openassetio-python-bridge` is not a Python module per se, its
dependencies are similar, in that we expect Python to be provided
by the consumer of the library, rather than linking to it explicitly.

Also ensure pybind11 requirement is not included in the generated CMake
target file(s) in the install tree.

Signed-off-by: David Feltell <david.feltell@foundry.com>