[Tests] Exclude examples/host from ctest
The simple resolver example host requires BAL for testing (so is more an
integration test). We are shortly splitting BAL out into it's own repo,
and so the processing of running those tests needs to change. This is
going to happen shortly anyway, as part of our source tree refactor
(see #628), so for now, well just skip running those tests here.

Signed-off-by: Tom Cowland <tom@foundry.com>