[Resources] Require external BAL installation for simpleResolver
BAL is shortly to be moved out of the main OpenAssetIO repository, and
so we can't assume it will be in some relative path by default.

We plan to add test coverage via an integration CI workfow, and later
provide an umbrella repository that aids running fully integrated tests
that cover MediaCreation/BAL/etc. as part of the TDD loop.

This commit updatest the documentation and tests to make use of external
dependencies.

Signed-off-by: Tom Cowland <tom@foundry.com>