[Build] Remove Python version override
Part of #654. Overriding the version of Python for `find_package` is
something we do not have a use case for, and is brittle. The
version specifier is only a minimum version, so does not help pin to
a specific version. The more common issue than version mismatch is
discovering the wrong Python on the system, for which we have to supply
additional hints (e.g. `Python_EXECUTABLE`), which supersedes any
version string anyway.

Signed-off-by: David Feltell <david.feltell@foundry.com>