[PluginSystem] Support entry point based plugin discovery
Extends the PythonPluginSytem with an additional method
`scan_entry_points` that allows plugins to be loaded using `importlib`
based entry point discovery.

`importlib.metadata` isn't available until Python 3.8, and the stable
form of its API doesn't appear until 3.10. As such, due to our need to
support 3.7-3.9 for at least several years to come, we use the
maintained backport version `importlib_metadata` until we are at a
minimum supported version of 3.10, where we can move to the native
`importlib.metadata` version.

We attempt to align env var names with OpenTimelineIO's precedent to
hopefully make things familiar and more straight forward - given both
technologies are likely to be in use in the same environment.

It is noted that the test coverage for `PythonPluginSystem` is already a
little lacking (in terms of error case handling for malformed plugins),
this will be addressed in a subsequent task.

Closes #762

Signed-off-by: Tom Cowland <tom@foundry.com>