[Docs] Add Sandboxed Builds documentation
In service of documenting the build process for the openassetio-build
container, create a new doc file relating to sandbox build environments.

This has been placed in the root of the repository for the moment,
along with BUILDING. This is in anticipation of it moving to a `docs`
folder in the project restructure.

Issue #716

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>