[Build] Add Testing.cmake module
Part of #655. In preparation for having CTest tests defined outside the
`tests` directory, we need to define the helpers globally, rather than
scoped only under the `tests` directory.

Since CMake's `include()` essentially copy-pastes the content of the
file, we must ensure the new `Testing` module is only `include`d once,
to avoid redefining global variables/targets. This then means we must
also ensure the file is included from a top-level project directory, so
that the global variables/targets are available to all subdirectories.

The use (or not) of `test-venv` is now decentralised, so add a new
function that will conditionally add `openassetio-python-venv` as a
fixture, if enabled.

Signed-off-by: David Feltell <david.feltell@foundry.com>