[Build] Move setup.py under openassetio-python
Closes #728 and follows on from the project restructure in #655. During
the project restructure we were not sure if wheel builds could continue
to work if we moved setup.py; and had an open question of whether we
wanted to move setup.py anyway, since it helps discoverability if we
locate setup.py at the root of the project.

After a little investigation, the main headache in the build process is
resolved by simply upgrading pip. Later pip versions build in-tree
(rather than copying the package to a temporary directory), so that
relative paths (e.g. to `CMakeLists.txt`) can be resolved.

In addition, we decided that on balance it's better to colocate Python
artifacts together under the `openassetio-python` component directory.
This has the added benefit of further disambiguating the project
at-a-glance to make it clear that this is a CMake C++ project primarily,
with an optional Python component.

Signed-off-by: David Feltell <david.feltell@foundry.com>