[Tests] Split out "bridge" tests from "internal"
Part of #739. We were combining the "bridge" tests of the
`openassetio-python-bridge` with the "module" tests of Python bindings
helpers in `openassetio-python-module` and, somewhat arbitrarily,
grouping them under an "internal" heading.

These tests all relied on running under `pytest` and using an
`_openassetio_test` Python extension module to provide C++ helpers. This
makes sense for the "module" tests (`PyRetainingSharedPtr`) but was a
quick-and-dirty solution for the "bridge" tests
(`createPythonPluginSystemManagerImplementationFactory`).

So move the "bridge" tests out into their own directory with a C++
CTest that embeds a Python interpreter. This is much more representative
of how we expect the `openassetio-python-bridge` library to be used.

Signed-off-by: David Feltell <david.feltell@foundry.com>