[Core] Remove media creation specific traits
Removes trait definitions from the core repository, so they can be
refined/explored in a more domain-specific space. These will be added
to the MediaCreation repo instead.

We had previously considered the 'Managed' Trait to be a universal
truth. However, explorations into other sectors have revealed that this
is perhaps making assumptions. Other sectors may "manage" data with
different granularities, vs an arbitrary "yes or no" approach.

This commit also takes a first step at improving some of the docs, that
are now overly specific. The topic of "consult your domain-specific
traits package for instructions" message needs weaving further into
other areas of the docs, and so will be picked up in more detail in a
follow up task.

This also updates the WillManagePathTrait to the alternate
ResolvesFutureEntitiesTrait addressing a long-standing issue that the
old formulation was too closely tied the old "primary string" based
workflow (vs the new composed traits approach).

Signed-off-by: Tom Cowland <tom@foundry.com>