[Core] Acquire GIL before C++ imports Python plugin system
Fixes #797. If the host thread does not hold the GIL when
`createPythonPluginSystemManagerImplementationFactory` is called, then
segfaults abound when trying to `import` the Python plugin system via
pybind11.

So ensure the GIL is acquired first. pybind11's `gil_scoped_acquire`
is clever enough to restore the state to whatever it was before, upon
function exit.

Signed-off-by: David Feltell <david.feltell@foundry.com>