[Core] Fix PyObject cleanup if interpreter gone
The `PyRetainingSharedPtr` mechanism in OpenAssetIO allows a C++
`shared_ptr` chain to keep a `PyObject` alive, even if that object is
not currently visible from the Python interpreter. This allows e.g. a
C++ host to hold a reference to (and call) a Python manager plugin.

A side effect of this mechanism is that it is possible for a C++ object
to keep a `PyObject` alive beyond the lifetime of the Python
interpreter. Then, when the C++ object is finally destroyed (e.g. on
application exit), attempting to clean up the `PyObject` is a
use-after-free error, causing segfaults or hangs.

So ensure cleanup of the `PyObject` is trivial, in the case that the
Python interpreter has been shut down, by not trying to acquire the GIL
and by `release`ing the pybind11 handle.

Signed-off-by: David Feltell <david.feltell@foundry.com>