[Core] Add ability to create a new manager for each harness test case
The `openassetio.test.manager` harness used to create a single manager
for the entire test run, reusing it for each case. This was mainly for
performance reasons.

This does however, preclude the ability to test manager initialization
behaviour.

This adds a class variable that instructs the harness to create a new,
uninitialized manager for each case instead.

Part of OpenAssetIO/OpenAssetIO-Manager-BAL#26.

Signed-off-by: Tom Cowland <tom@foundry.com>