[Lint] Conform to cmake-lint
In order to avoid cmake-errors and redundant style discussions, we're
integrating cmake-lint, and conforming all our cmake files to it.

The style document, cmake-format.yaml, is the default settings, save for
the line length and tab width, which are set to our standards, 99 and 4.

Choosing to conform to the defaults is simply the path of least
resistance. Our private variables were snake_cased before, but are now
_snake_cased with the underscore in front, which isn't so big a change.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>