[CI] Fix CI for latest version of `ncurses` recipe
We must pin the `ncurses` package (a dependency of `cpython`) to the
previous recipe revision, since the latest supplies incorrect keyword
args to the `patch` function (possibly a Conan v2 -specific change).

If we need to pin a package to a specific Conan recipe revision, then we
need to explicitly opt-in to this functionality by setting
`general.revisions_enabled=False` in conan.conf. Frustratingly, a
`self.requires(...)` in conanfile.py doesn't complain if you try to use
a pinned revision, and will silently use the newest recipe. This is in
contrast to the CLI, which does at least complain.

Signed-off-by: David Feltell <david.feltell@foundry.com>