[Docs] Add additional context about API use
We've seen more and more people using simpleResolver.py as a 'getting
started' example for OpenAssetIO. This is great, but due to the nature
of it being a naive resolver, it's a bad example in terms of key
best-practices. This commit attempts to call out where this is the case,
and also provide additional context/tips if the reader happens to be a
host integrator looking for guidance.

Signed-off-by: Tom Cowland <tom@foundry.com>