[Examples] Remove custom locale from simpleResolver
We used to encourage people to make on-the-stack traits/specs. As we
have matured our understanding how this should all work as an ecosystem,
this has become an antipattern. A stack-lifetime trait or specification
is useless outside that particular code base. The whole reason for these
'view classes' is to ensure consistency between different parts of a
pipeline, so a host should either:

- Author a TraitGen YAML file, so they can be used downstream/upstream.
- Be honest and just use string literals and/or export constants.

Part of #835

Signed-off-by: Tom Cowland <tom@foundry.com>