[Core] Add resolve convenience overloads
Closes #849. Previously, the only interface to `resolve` assumed a
batch of entity references with success/error callbacks executed for
each element. This is somewhat clunky when only resolving a single
reference in particular, and requires significant boilerplate.

So add convenience overloads that present the API in a friendlier and
more versatile manner.

As resolve conveniences in C++ and Python are going to be somewhat
distinct, add C++ tests to verify behaviour of `variant` and exception
error modes.

Signed-off-by: David Feltell <david.feltell@foundry.com>