[Lint] Fix clang-tidy v15 warnings
Although we have yet to move to the newer clang-tidy on CI, we aim to
support devs using a newer clang-tidy than v12.

Interestingly, fixing warnings in newer clang-tidy can make older (v12)
clang-tidy unhappy.  Apparently changing a variable to `const` causes
`bugprone-infinite-loop` to trigger on Catch2 `CHECK` macros. So add
a suppression for this.

Signed-off-by: David Feltell <david.feltell@foundry.com>