[Core] Deep-copy locale in `Manager.createChildContext`
`createChildContext` is meant to produce a new context that is
associated with the parent. Speicifcally, for when you wish to modify an
existing context in paralell. Sharing a locale (as we were only copying
the ptr) meant that modifing the locale in one would update the locale
in another. This is confusing and undesirable.

Now the locale `TraitsData` is deep-copied.

Closes #896

Signed-off-by: Tom Cowland <tom@foundry.com>