[Core] Remove test harness locale view classes
Part of #835. `TraitBase` and `SpecificationBase` are due to be removed
from the core repo, with trait and view classes instead generated using
the OpenAssetIO-TraitGen project.

So we must remove the remaining trait and specification view classes
from the core. The only remaining example of this is the locale
specifications/traits for the manager test harness.

We still need a locale for the test harness, but very little of it is
actually in use, so strip it down to the basics and construct a locale
`TraitsData` using the low-level API.  I.e. a single
`"openassetio:test.locale"` trait with a `"case"` property.

Signed-off-by: David Feltell <david.feltell@foundry.com>