[Core] Value types in batch callbacks
[#858]
It is normal that a user receiving a value via a callback will be
expected to take ownership of the value, and may want to store
it for later. This has happened already for us, in our internal
conveniences.

This is problematic to do with a const-ref, causing either awkward
reference holding or inadvertant copies. Change to value semantics
to make ownership clear and empower the users to use these objects
optimally.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>