[Core] Refactor `getRelatedReferences`
The old "two lists" formulation was hard to grok, and difficult to
implement effectively. Splitting into two discrete methods better
outlines the two potential batch axis.

This new approach makes it clear the two well-known divergent use cases
and allows for a simpler code path in each - without the need to
interrogate the input arrays to work out which pattern is being used
(one-to-many or many-to-one).

Naming this was hard, after assorted rounds, this option emerged as a
'best of a bad bunch', as it is at least consistent with future
'addRelationship' and 'removeRelationship' methods, along with our more
general noun/verb conventions. There is certainly ambiguity as the
singular/plural difference is subtle, but it's the best we could come up
with.

These new methods also use the batch callback pattern established with
the other batch API signatures.

In addition, we have always documented that supporting relationship
queries is optional for a manager (though highly recommended). Despite
this the API methods were marked as being abstract. We only got away
with this as since the move to Python 3, ABC enforcement wasn't working.

There are now default implementations of `getWithRelationship` and
`getWithRelationships` that simply return an empty list for each
element of the batch.

NB. This commit disables BAL integration tests due to the API break.

Closes #847
Closes #919

Signed-off-by: Tom Cowland <tom@foundry.com>