[Core] Refactor `getRelatedReferences`
The old "two lists" formulation was hard to grok, and difficult to
implement effectively. Splitting into two discrete methods better
outlines the two potential batch axis.

This new approach makes it clear the two well-known divergent use cases
and allows for a simpler code path in each - without the need to
interrogate the input arrays to work out which pattern is being used
(one-to-many or many-to-one).

Naming this was hard, after assorted rounds, this option emerged as a
'best of a bad bunch', as it is at least consitent with future
'addRelationship' and 'removeRelationship' methods, along with our more
general noun/verb conventions. There is certainly ambiguity as the
singular/plural difference is subtle, but it's the best we could come up
with.

Closes #847

Signed-off-by: Tom Cowland <tom@foundry.com>