[Core] Make `getWithRelationship[s]` optional
We have always documented that supporting relationship queries is
optional for a manager (though highly recommended). Despite this the
API methods were marked as being abstract. We only got away with this
as since the move to Python 3, ABC enforcement wasn't working.

This adds default implementations of `getWithRelationship` and
`getWithRelationships` that simply returns an empty list for each
element of the batch.

Part of #847

Signed-off-by: Tom Cowland <tom@foundry.com>

Signed-off-by: Tom Cowland <tom@foundry.com>