[Core] defaultManagerForInterface configPath arg
Closes #937.

There are many scenarios where end-users do not have ready means of
configuring the process environment (eg: double-clicking on an icon),
and a host may way to provide alternate means of determining the path to
the default config file.

Add a new overload of `defaultManagerForInterface` that uses a
`configPath` argument, rather than an environment variable, to determine
the location of the TOML config file.

Signed-off-by: David Feltell <david.feltell@foundry.com>