[Core] Add ${config_dir} config TOML interpolation
Closes #804.

It is common in test scenarios to create portable, self-contained
manager states using BAL.

At the moment, this requires relative paths in the TOML file, and so
becomes working directory sensitive. This makes it fragile.

So add support for the substitution in TOML string values of any
`${config_dir}` tokens with the absolute directory to the TOML config
file.

This is currently done in the most basic way possible, by direct string
find-and-replace, rather than a more complete solution supporting e.g.
environment variable interpolation. Such improvement is for future work
if/when the need ever arises.

Signed-off-by: David Feltell <david.feltell@foundry.com>