[Core] Migrate getWithRelationship(s) to C++
Part of #913. This is required so that C++ hosts can call these methods,
and so that C++ managers can implement and override them.

One small API break is to make the `resultTraitSet` argument
non-optional in `ManagerInterface`. Clang-Tidy doesn't like default
values for arguments of `virtual` member functions - presumably because
default arguments are not inherited. But since it's always going to be
wrapped by a `Manager`, we can make the argument optional there instead,
which is arguably the right thing to do anyway.

Signed-off-by: David Feltell <david.feltell@foundry.com>