[Docs] Add expand-to-content width
Changes the width of the main content div to by default, always
be 1100px. This is effectively the same as our prior max-width in
almost every case, as we always have prose that can push the
width to that limit and cause a wrap.

The `width: min-content` flag, in combination with using min rather
than max, overrides this in cases where the contents are fixed width
and cannot wrap (like method sig tables,) and allows the frame to be
pushed out to support its minimum width.

However, the problem with this approach would be that smaller
frame sizes that _could_ support non-fixed widths, (ie, all of our
pages that are just prose and can wrap very tightly), would no
longer wrap below 1100px. To that end, use a media query to
remove the minimum width when the frame gets that narrow.

I think this is a good compromise, that keeps our nice column view
whilst also allowing small screens/split windows to get wrapped
text, and allowing our unwrappable elements to push the frame
size wider.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>