[Core] Return None when castToPyObject of nullptr
Closes #988. In real-world use cases, it is useful to get a `None` value
from an uninitialised `shared_ptr` when passed to `castToPyObject`,
rather than that being an error case. E.g. testing whether a manager
(`ManagerPtr`) has been created or not.

Note that the converse was already the case, i.e. `castFromPyObject`,
when given `None`, would return a `nullptr`, but we didn't have tests
for that case (until now).

The behaviour when passing a `nullptr` `PyObject` to `castFromPyObject`
has not been changed - we will still throw an exception in that case.
This seems reasonable, since the alternative is a segfault.

Signed-off-by: David Feltell <david.feltell@foundry.com>