[Core] isEntityReferenceString prefix optimisation
Closes #566. When a Python based manager implementation is used, the
cost of acquiring the GIL and entering Python in order to invoke a
simple string prefix check can be prohibitive.

We have several docstrings that refer to the
`kInfoKey_EntityReferencesMatchPrefix` optimisation/convenience of
`isEntityReferenceString`. However, it was not supported, making the
docs invalid.

So add support for using this optional optimisation feature.

Signed-off-by: David Feltell <david.feltell@foundry.com>