[Core] Add close method to PagerInterface
Add `close` method to pager interface, called upon destruction of pager

There were various subtle issues with tying the notification of the end
of the page query to the destruction of the pager interface.
Aside from relying on lifetime like that not being very pythonic
(having to override del,) it creates an implicit usage assumption
that the PagerInterface is not declared in a greater scope than
the pager. Not to mention that the prior design encourages
throwing in destructors moreso than this new one.

Adding a close method binds the cleanup entry hook directly
to the lifetime of the Pager, rather than the PagerInterface.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>