[Docs] Breaking change release note for `constants` import
In #998 the `openassetio.constants` pure-Python module was migrated to a
CPython submodule (`_openassetio.constants`), which has the unfortunate
side-effect of no longer allowing the `import` of individual elements
inside the module (i.e. `from constants import <something>`).

So update the release notes to call that out as an additional breaking
change.

This was detected (and fixed) in BAL in
OpenAssetIO/OpenAssetIO-Manager-BAL#55

Signed-off-by: David Feltell <david.feltell@foundry.com>