[Core] Increase flexibility of publish workflow
Closes #1029 #1029

A well-behaved host should give the manager as much information as it
can at `preflight` time to have the best chance of the manager
supporting intermediate `resolve`-for-`kWrite` and the final `register`
call.

That is, information that the host "owns" should be provided to the
manager on `preflight` if it's available.

`register` validated that all `TraitsData`s for a batch had the same
trait set. We largely "guessed" that this constraint would make it
easier for a manager to feasibly batch any given API call to its
back-end. It removed the need for the manager to determine if all the
trait sets are the same.

This is overly restrictive in the case where a host legitimately wishes
to publish several entities with different traits, as it forces each of
these to make individual API calls. As such, the manager has no
opportunity to batch the back-end requests even if it could do so, as
they are discrete API calls.

So redesign the publishing workflow such that `preflight` takes a
`TraitsData` per batch element, instead of a trait set for the whole
batch, and remove the restriction that all the `TraitsData`s in a batch
must have the same trait set.

Add a new `BatchElementError` code of `kInvalidPreflightHint`, to be
returned from `preflight` for elements where a `TraitsData` is
insufficient or invalid for servicing subsequent `resolve`/`register`
calls.

Signed-off-by: David Feltell <david.feltell@foundry.com>