[Python] Increase verbosity of API compliance suite
It's very easy for manager plugin authors to accidentally miss defining
a fixture, causing a test to be skipped erroneously. With the default
report, it is difficult to discern which tests were skipped.

So increase the verbosity, so that all tests are printed along with
their status, including the reason given for skipping a test.

Signed-off-by: David Feltell <david.feltell@foundry.com>