[Docs] Disambiguate glossary from methods
Part of #124. When writing docstrings for #1037, it is useful to link to
the `preflight` glossary from within `Manager` and `ManagerInterface`
classes. However, `@ref preflight` links to the method in the class, not
the glossary.

So introduce a partial fix, handling just `resolve`, `preflight` and
`register` to prefix their glossary IDs with "glossary_". Note that
"glossary:" (as specified in #124) is not accepted by Doxygen as a valid
prefix.

As it turned out `register` wasn't ambiguous, because the method has a
trailing `_`, but prefixing the glossary ID revealed a few cases where
we have likely accidentally linked the wrong thing.

Some common sense has been used to figure out whether existing
references should refer to the glossary or the method, but this was not
pondered in detail, so we should probably revisit this in subsequent
work.

Signed-off-by: David Feltell <david.feltell@foundry.com>