[Test] Alter which fixture skips relation tests
#1022
Alter the skip fixture in certain getWithRelation tests away from
common shared ones  to the `a_relationship_trait_set`, which is
less likely to be defined in a shared context.

This is helpful because relationships are an optional feature, so it
is likely managers will want to skip all the relationship tests as a
group, and this makes that more or less the default behaviour without
thinking about it.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>