[Core] Extract `Access` from Context
Closes  #1056. Follows on from DR019 (#1051).

Observations of real-world integration efforts have highlighted just how
easy it is to miss-manage the context, resulting in hard to spot
programming/logic errors. We need to make the API as robust and easy to
learn as possible (it's hard enough as it is).

So extract `access` from the `Context` class, making it a first-class
argument instead.

Split access constants per-workflow, using `enum class`, so that it is
not possible to specify an unsupported access category to API methods.

However, ensure the enum values across workflow-specific enums are
consistent. Hence add a top-level C-style `enum` that dictates the
shared values. This can be extracted into a C-compatible header (and
appropriately re-namespaced) when we implement the C API. Prior similar
work used macros to achieve the same C compatibility layer (errors.h),
but these probably should have been enums too.

Signed-off-by: David Feltell <david.feltell@foundry.com>